requirements.txt
Django==2.1.1
pytz==2018.5
asn1crypto==0.24.0
backcall==0.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cli-helpers==1.0.2
click==6.7
colorama==0.3.9
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
decorator==4.3.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
ipython==6.4.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.12.0
Jinja2==2.10
jsonfield==2.0.2
MarkupSafe==1.0
mongoengine==0.15.0
mycli==1.17.0
mysqlclient==1.3.12
numpy==1.14.5
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
pymongo==3.6.1
PyMySQL==0.8.1
requests==2.19.1
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
tabulate==0.8.2
terminaltables==3.1.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
